requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.9.1
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.14.0
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
appdirs==1.4.4
apprise==1.9.1
arrow==1.3.0
asgi-lifespan==2.1.0
asttokens==3.0.0
asyncpg==0.30.0
asyncssh==2.19.0
atpublic==5.0
attrs==24.3.0
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
binaryornot==0.4.4
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.0
codespell==2.3.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
ConfigArgParse==1.7
configobj==5.0.9
contourpy==1.3.1
cookiecutter==2.6.0
coolname==2.2.0
coverage==7.6.10
croniter==5.0.1
cryptography==43.0.3
cycler==0.12.1
datasets==3.2.0
dateparser==1.2.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
deprecation==2.1.0
dictdiffer==0.9.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
docker==7.1.0
docker-pycreds==0.4.0
docopt==0.6.2
dpath==2.2.0
dulwich==0.22.7
dvc==3.58.0
dvc-data==3.16.7
dvc-gdrive==3.0.1
dvc-gs==3.0.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
dynaconf==3.2.6
entrypoints==0.4
evidently==0.5.1
exceptiongroup==1.2.2
executing==2.1.0
Faker==33.3.0
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
flatbuffers==24.12.23
flatten-dict==0.4.2
flufl.lock==8.1.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
funcy==2.0
gcsfs==2024.9.0.post1
gdown==5.2.0
gevent==24.11.1
geventhttpclient==2.3.3
ghp-import==2.1.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grandalf==0.8
graphviz==0.20.3
greenlet==3.1.1
griffe==1.5.4
grpcio==1.68.1
gto==1.7.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.0
humanfriendly==10.0
humanize==4.11.0
hydra-core==1.3.2
hyperframe==6.0.1
identify==2.6.4
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
invoke==2.2.0
ipykernel==6.29.5
ipython==8.12.3
iterative-telemetry==0.0.9
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jinja2-humanize-extension==0.4.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kagglehub==0.3.6
kiwisolver==1.4.8
kombu==5.4.2
lightning-utilities==0.11.9
litestar==2.13.0
litestar-htmx==0.4.1
locust==2.32.5
Mako==1.3.8
Markdown==3.7
markdown-exec==1.10.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
mistune==3.1.0
mkdocs==1.6.1
mkdocs-exclude==1.0.2
mkdocs-get-deps==0.2.0
mkdocs-git-revision-date-localized-plugin==1.3.0
mkdocs-glightbox==0.4.0
mkdocs-material==9.5.49
mkdocs-material-extensions==1.3.1
mkdocs-same-dir==0.1.3
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.16
mypy==1.14.1
mypy-extensions==1.0.0
narwhals==1.20.1
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
neural_compressor==3.2
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.20.1
opencv-python-headless==4.10.0.84
opentelemetry-api==1.29.0
optuna==4.1.0
orjson==3.10.13
packaging==24.2
paginate==0.5.7
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
pendulum==3.0.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
polyfactory==2.18.1
pre_commit==4.0.1
prefect==3.1.11
prettytable==3.12.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.1
ptflops==0.7.4
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.4
pydantic-extra-types==2.10.1
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.3
PyDrive2==1.21.3
pygit2==1.16.0
Pygments==2.18.0
pygments-ansi-color==0.3.0
pygtrie==2.5.0
pymdown-extensions==10.13
pyOpenSSL==24.2.1
pyparsing==3.2.1
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-socks==2.6.1
pytorch-lightning==2.5.0.post0
pytz==2024.2
PyYAML==6.0.2
pyyaml_env_tag==0.1
pyzmq==26.2.0
readchar==4.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-click==1.8.5
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.18.8
ruamel.yaml.clib==0.2.12
ruff==0.8.4
safetensors==0.5.0
schema==0.7.7
scikit-learn==1.5.2
scipy==1.15.0
scmrepo==3.3.9
semver==3.0.2
sentry-sdk==2.19.2
setproctitle==1.3.4
setuptools==75.6.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.17.0
smmap==5.0.2
snakeviz==2.2.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqltrie==0.11.1
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.4
streamlit==1.41.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
text-unidecode==1.3
threadpoolctl==3.5.0
time-machine==2.16.0
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
torch==2.5.1
torch-tb-profiler==0.4.3
torchmetrics==1.6.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
triton==3.1.0
typer==0.13.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.3.0
uuid6==2024.7.10
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
virtualenv==20.28.1
voluptuous==0.15.2
wandb==0.19.1
watchdog==6.0.0
watchfiles==1.0.3
wcwidth==0.2.13
webencodings==0.5.1
websockets==13.1
Werkzeug==3.1.3
wrapt==1.17.0
xxhash==3.5.0
yarg==0.1.9
yarl==1.18.3
zc.lockfile==3.0.post1
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
